Corrected the "0 index" issue.
Couldn't correct the rest if there is anything to correct. It seems to
be running, but a single iteration is taking up to a very long time (I
didn't finish it because I couldn't afford waiting for it), and it is
only for grayscale images. Basically I was useless :dancers: